refactor: move safe_get to utils/http.py
